updates to db to acess the id

+++ android-app-helloworld-master/app/src/main/java/com/navisens/demo/android_app_helloworld/SelectPathActivity.java
@@ -82,7 +82,7 @@ public class SelectPathActivity extends AppCompatActivity {
         }
     }
 
-    private void startNewActivity(Class activity, int pid) {
+    private void startNewActivity(Class activity, long pid) {
         Intent intent = new Intent(this, activity);
         intent.putExtra("currentPath", pid);
         startActivity(intent);
