+++ android-app-helloworld-master/app/src/main/java/com/navisens/demo/android_app_helloworld/database_obj/PathPointDao.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Dao
 public interface PathPointDao {
     @Query ("SELECT * FROM points WHERE path_id = :p ORDER BY path_point_id ASC")
-    public List<PathPoint> getByPathId(int p);
+    public List<PathPoint> getByPathId(long p);
 
     @Insert
     public void addPathPoints(List<PathPoint> p);
@@ -19,7 +19,7 @@ public interface PathPointDao {
     public void addPoint(PathPoint p);
 
     @Query("DELETE FROM points WHERE path_id = :path_id")
-    public void deleteByPathId(int path_id);
+    public void deleteByPathId(long path_id);
 
     @Query("DELETE FROM points")
     public void deleteAll();
